+++ src/app/mod.rs
@@ -7,19 +7,103 @@ mod flake;
 mod health;
 mod info;
 
-use leptos::*;
-use leptos_extra::{
-    query::{self},
-    signal::{provide_signal, SignalWithResult},
-};
-use leptos_meta::*;
-use leptos_query::*;
-use leptos_router::*;
-use nix_rs::{command::Refresh, flake::url::FlakeUrl};
+use dioxus::prelude::*;
+use dioxus_router::prelude::*;
 
-use crate::{app::flake::*, app::health::*, app::info::*, widget::*};
+use crate::widget::*;
+
+#[derive(Routable, PartialEq, Debug, Clone)]
+enum Route {
+    #[route("/")]
+    Dashboard {},
+    #[route("/about")]
+    About {},
+}
 
 /// Main frontend application container
+pub fn App(cx: Scope) -> Element {
+    // TODO: per-route title
+    render! {
+        body {
+            div { class: "flex justify-center w-full min-h-screen bg-center bg-cover bg-base-200",
+                div { class: "flex flex-col items-stretch mx-auto sm:container sm:max-w-screen-md",
+                    main { class: "flex flex-col px-2 mb-8 space-y-3 text-center",
+                        Nav {}
+                        p { "It is WIP" }
+                        ul { li { "Tailwind works" } }
+                        img { src: "images/128x128.png" }
+                    }
+                }
+            }
+        }
+    }
+}
+
+// Home page
+fn Dashboard(cx: Scope) -> Element {
+    tracing::debug!("Rendering Dashboard page");
+    /* let result = query::use_server_query(cx, || (), get_nix_health);
+    let data = result.data;
+    let healthy = Signal::derive(cx, move || {
+        data.with_result(|checks| checks.iter().all(|check| check.result.green()))
+    });
+    */
+    // A Card component
+    #[inline_props]
+    fn Card<'a>(cx: Scope, href: &'static str, children: Element<'a>) -> Element<'a> {
+        render! {
+            a {
+                href: "{href}",
+                class: "flex items-center justify-center w-48 h-48 p-2 m-2 border-2 rounded-lg shadow border-base-400 active:shadow-none bg-base-100 hover:bg-primary-200",
+                span { class: "text-3xl text-base-800", children }
+            }
+        }
+    }
+    render! {
+        h1 { class: "text-5xl font-bold", "Dashboard" }
+        div { id: "cards", class: "flex flex-row flex-wrap",
+            Card { href: "/info", "Nix Info ℹ️" }
+            Card { href: "/flake", "Flake Overview ❄️️" }
+        }
+    }
+}
+
+/// Navigation bar
+///
+/// TODO Switch to breadcrumbs, as it simplifes the design overall.
+fn Nav(cx: Scope) -> Element {
+    let class = "px-3 py-2";
+    render! {
+        nav { class: "flex flex-row w-full mb-8 text-white md:rounded-b bg-primary-800",
+            a { href: "/", class: class, "Dashboard" }
+            a { href: "/flake", class: class, "Flake" }
+            a { href: "/health", class: class, "Nix Health" }
+            a { href: "/info", class: class, "Nix Info" }
+            a { href: "/about", class: class, "About" }
+            div { class: "flex-grow font-bold text-end {class}", "🌍 nix-browser" }
+        }
+    }
+}
+
+/// About page
+fn About(cx: Scope) -> Element {
+    render! {
+        h1 { "About" }
+        h1 { class: "text-5xl font-bold", "About" }
+        p {
+            "nix-browser is still work in progress. Track its development "
+            a {
+                href: "https://github.com/juspay/nix-browser",
+                class: "underline text-primary-500 hover:no-underline",
+                rel: "external",
+                target: "_blank",
+                "on Github"
+            }
+        }
+    }
+}
+
+/*
 #[component]
 pub fn App(cx: Scope) -> impl IntoView {
     provide_meta_context(cx);
@@ -60,87 +144,4 @@ pub fn App(cx: Scope) -> impl IntoView {
         </Router>
     }
 }
-
-/// Navigation bar
-///
-/// TODO Switch to breadcrumbs, as it simplifes the design overall.
-#[component]
-fn Nav(cx: Scope) -> impl IntoView {
-    let class = "px-3 py-2";
-    view! { cx,
-        <nav class="flex flex-row w-full mb-8 text-white md:rounded-b bg-primary-800">
-            <A exact=true href="/" class=class>
-                "Dashboard"
-            </A>
-            <A exact=false href="/flake" class=class>
-                "Flake"
-            </A>
-            <A exact=true href="/health" class=class>
-                "Nix Health"
-            </A>
-            <A exact=true href="/info" class=class>
-                "Nix Info"
-            </A>
-            <A exact=true href="/about" class=class>
-                "About"
-            </A>
-            <div class=format!("flex-grow font-bold text-end {}", class)>"🌍 nix-browser"</div>
-        </nav>
-    }
-}
-
-/// Home page
-#[component]
-fn Dashboard(cx: Scope) -> impl IntoView {
-    tracing::debug!("Rendering Dashboard page");
-    let result = query::use_server_query(cx, || (), get_nix_health);
-    let data = result.data;
-    let healthy = Signal::derive(cx, move || {
-        data.with_result(|checks| checks.iter().all(|check| check.result.green()))
-    });
-    // A Card component
-    #[component]
-    fn Card(cx: Scope, href: &'static str, children: Children) -> impl IntoView {
-        view! { cx,
-            <A
-                href=href
-                class="flex items-center justify-center w-48 h-48 p-2 m-2 border-2 rounded-lg shadow border-base-400 active:shadow-none bg-base-100 hover:bg-primary-200"
-            >
-                <span class="text-3xl text-base-800">{children(cx)}</span>
-            </A>
-        }
-    }
-    view! { cx,
-        <Title text="Dashboard"/>
-        <h1 class="text-5xl font-bold">"Dashboard"</h1>
-        <div id="cards" class="flex flex-row flex-wrap">
-            <SuspenseWithErrorHandling>
-                <Card href="/health">
-                    "Nix Health Check "
-                    {move || {
-                        healthy
-                            .with_result(move |green| {
-                                view! { cx, <CheckResultSummaryView green=*green/> }
-                            })
-                    }}
-
-                </Card>
-            </SuspenseWithErrorHandling>
-            <Card href="/info">"Nix Info ℹ️"</Card>
-            <Card href="/flake">"Flake Overview ❄️️"</Card>
-        </div>
-    }
-}
-
-/// About page
-#[component]
-fn About(cx: Scope) -> impl IntoView {
-    view! { cx,
-        <Title text="About"/>
-        <h1 class="text-5xl font-bold">"About"</h1>
-        <p>
-            "nix-browser is still work in progress. Track its development "
-            <LinkExternal link="https://github.com/juspay/nix-browser" text="on Github"/>
-        </p>
-    }
-}
+*/
